Implemented state 2 of door passing

+++ door_passing/include/door_passing/door_passing.h
@@ -33,7 +33,7 @@ private:
     double turn_range_;
     double distance_to_door_;
     double distance_after_passing_;
-    double state_;
+    int state_;
 };
 
 #endif
